Added authorization and authentication

+++ backend/src/modules/contact/contact.routes.ts
@@ -9,11 +9,15 @@ import {
     createContactSchema,
     updateContactSchema,
 } from "./contact.schema";
+import { authenticate } from "../../middleware/auth.middleware";
+import { authorize } from "../../middleware/authorize.middleware";
 
 const router = Router();
 
 router.post(
     "/",
+    authenticate,
+    authorize("ADMIN"),
     validate(createContactSchema),
     asyncHandler((req, res) =>
         contactController.create(req, res)
@@ -22,6 +26,8 @@ router.post(
 
 router.get(
     "/",
+    authenticate,
+    authorize("ADMIN"),
     asyncHandler((req, res) =>
         contactController.findAll(req, res)
     )
@@ -29,6 +35,8 @@ router.get(
 
 router.get(
     "/:id",
+    authenticate,
+    authorize("ADMIN"),
     asyncHandler((req, res) =>
         contactController.findById(req, res)
     )
@@ -36,6 +44,8 @@ router.get(
 
 router.patch(
     "/:id",
+    authenticate,
+    authorize("ADMIN"),
     validate(updateContactSchema),
     asyncHandler((req, res) =>
         contactController.update(req, res)
@@ -44,6 +54,8 @@ router.patch(
 
 router.delete(
     "/:id",
+    authenticate,
+    authorize("ADMIN"),
     asyncHandler((req, res) =>
         contactController.delete(req, res)
     )
