+++ backend/src/modules/document/dosument.routes.ts
@@ -9,11 +9,15 @@ import {
     createDocumentSchema,
     updateDocumentSchema,
 } from "./document.schema";
+import { authenticate } from "../../middleware/auth.middleware";
+import { authorize } from "../../middleware/authorize.middleware";
 
 const router = Router();
 
 router.post(
     "/",
+    authenticate,
+    authorize("ADMIN"),
     validate(createDocumentSchema),
     asyncHandler((req, res) =>
         documentController.create(req, res)
@@ -22,6 +26,8 @@ router.post(
 
 router.get(
     "/",
+    authenticate,
+    authorize("ADMIN"),
     asyncHandler((req, res) =>
         documentController.findAll(req, res)
     )
@@ -29,6 +35,8 @@ router.get(
 
 router.get(
     "/:id",
+    authenticate,
+    authorize("ADMIN"),
     asyncHandler((req, res) =>
         documentController.findById(req, res)
     )
@@ -36,6 +44,8 @@ router.get(
 
 router.patch(
     "/:id",
+    authenticate,
+    authorize("ADMIN"),
     validate(updateDocumentSchema),
     asyncHandler((req, res) =>
         documentController.update(req, res)
@@ -44,6 +54,8 @@ router.patch(
 
 router.delete(
     "/:id",
+    authenticate,
+    authorize("ADMIN"),
     asyncHandler((req, res) =>
         documentController.delete(req, res)
     )
