+++ backend/src/modules/product-category/product-category.routes.ts
@@ -9,11 +9,15 @@ import {
 } from "./product-category.schema";
 
 import { productCategoryController } from ".";
+import { authenticate } from "../../middleware/auth.middleware";
+import { authorize } from "../../middleware/authorize.middleware";
 
 const router = Router();
 
 router.post(
     "/",
+    authenticate,
+    authorize("ADMIN"),
     validate(CreateProductCategorySchema),
     asyncHandler((req, res) =>
         productCategoryController.create(req, res)
@@ -22,6 +26,8 @@ router.post(
 
 router.get(
     "/",
+    authenticate,
+    authorize("ADMIN", "CUSTOMER"),
     asyncHandler((req, res) =>
         productCategoryController.findAll(req, res)
     )
@@ -29,6 +35,8 @@ router.get(
 
 router.get(
     "/:id",
+    authenticate,
+    authorize("ADMIN", "CUSTOMER"),
     asyncHandler((req, res) =>
         productCategoryController.findById(req, res)
     )
@@ -36,6 +44,8 @@ router.get(
 
 router.patch(
     "/:id",
+    authenticate,
+    authorize("ADMIN"),
     validate(UpdateProductCategorySchema),
     asyncHandler((req, res) =>
         productCategoryController.update(req, res)
@@ -44,6 +54,8 @@ router.patch(
 
 router.delete(
     "/:id",
+    authenticate,
+    authorize("ADMIN"),
     asyncHandler((req, res) =>
         productCategoryController.delete(req, res)
     )
